requirements.txt
aiogram==3.3.0
python-dotenv==1.0.1
sqlalchemy==2.0.27
alembic==1.13.1
pydantic==2.6.1
psycopg2-binary==2.9.9 